+++ mex/backend/types.py
@@ -2,6 +2,7 @@ from enum import Enum, EnumMeta, _EnumDict
 
 from pydantic import SecretStr
 
+from mex.backend.fields import REFERENCED_ENTITY_TYPES_BY_FIELD_BY_CLASS_NAME
 from mex.common.models import (
     EXTRACTED_MODEL_CLASSES_BY_NAME,
     MERGED_MODEL_CLASSES_BY_NAME,
@@ -73,3 +74,15 @@ class MergedType(Enum, metaclass=DynamicStrEnum):
     """Enumeration of possible types for merged items."""
 
     __names__ = list(MERGED_MODEL_CLASSES_BY_NAME)
+
+
+class ReferenceFieldName(Enum, metaclass=DynamicStrEnum):
+    """Enumeration of possible field names that contain references."""
+
+    __names__ = sorted(
+        {
+            dk
+            for v in REFERENCED_ENTITY_TYPES_BY_FIELD_BY_CLASS_NAME.values()
+            for dk in v
+        }
+    )