+++ CHANGELOG.md
@@ -10,12 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - add endpoints to get a single merged or single extracted item by id
+- add generic id filter to merge and preview endpoints
 
 ### Changes
 
 - rename `GraphConnector.exists_merged_item` to `exists_item`
 
 ### Deprecated
+
+- deprecate hadPrimarySource for merge and preview endpoints
 
 ### Removed